Converted Model, TimeSeriesModel and AnomalyDetectionModel back to interfaces
