+++ src/main/java/com/yahoo/egads/models/adm/AnomalyDetectionModel.java
@@ -12,23 +12,19 @@ import com.yahoo.egads.data.Anomaly;
 import com.yahoo.egads.data.TimeSeries;
 import com.yahoo.egads.data.Model;
 
-public abstract class AnomalyDetectionModel extends Model {
+public interface AnomalyDetectionModel extends Model {
     // methods ////////////////////////////////////////////////
 
     // returns the type of anomalies detected by the model
-    public abstract String getType();
+    public String getType();
 
-    public String getModelType() {
-    	return "Anomaly";
-    }
-    
     // tune the anomaly detection parameters based on the training data.
-    public abstract void tune(TimeSeries.DataSequence observedSeries,
+    public void tune(TimeSeries.DataSequence observedSeries,
             TimeSeries.DataSequence expectedSeries,
             Anomaly.IntervalSequence anomalySequence) throws Exception;
 
     // detect anomalies.
-    public abstract Anomaly.IntervalSequence detect(
+    public Anomaly.IntervalSequence detect(
             TimeSeries.DataSequence observedSeries,
             TimeSeries.DataSequence expectedSeries) throws Exception;
 }
